// Frontend/src/app/shared/interfaces/accommodation.dto.ts
export interface AccommodationApiDto {
  _id: unknown;
  checkInAt?: unknown;
  checkOutAt?: unknown;
  city?: unknown;
  postalCode?: unknown;
  hasPublicTransportNearby?: unknown;
  publicTransportTypes?: unknown;
  additionalDescription?: unknown;
  houseRules?: unknown;
  distanceToCourtsMeters?: unknown;
  address?: unknown;
  maxGuests?: unknown;
  photoUrl?: unknown;
  isDeleted?: unknown;
  ownerId?: unknown;
  tournamentId?: unknown;
  createdAt?: unknown;
  updatedAt?: unknown;
}